rename input header (h -> hpp)
