+++ src/silly_sprites/CMakeLists.txt
@@ -17,7 +17,7 @@ add_library(sillysprites
         opengl/glfw_context.cpp
         opengl/glfw_context.hpp
         types.hpp
-        opengl/input.h
+        opengl/input.hpp
         opengl/input.cpp
 )
